cb::mcbp::Response.getStatus() should return cb::mcbp::Status
To avoid having to do the conversion explicit

Reviewed-on: http://review.couchbase.org/99465
Tested-by: Build Bot <build@couchbase.com>

+++ include/mcbp/protocol/framebuilder.h
@@ -195,7 +195,7 @@ public:
     ResponseBuilder(cb::byte_buffer backing) : FrameBuilder<Response>(backing) {
     }
     void setStatus(Status status) {
-        getFrame()->setStatus(uint16_t(status));
+        getFrame()->setStatus(status);
     }
 };
